fix(server): preserve billing transaction boundaries
