+++ server/proliferate/server/billing/stripe_webhooks.py
@@ -7,10 +7,13 @@ import hmac
 import json
 import logging
 import time
+from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
 from datetime import UTC, datetime
-from typing import Any
+from typing import Any, Concatenate
 from uuid import UUID
+
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from proliferate.config import settings
 from proliferate.constants.billing import (
@@ -18,6 +21,7 @@ from proliferate.constants.billing import (
     PRO_PERIOD_GRANT_TYPE,
     REFILL_10H_GRANT_TYPE,
 )
+from proliferate.db import engine as db_engine
 from proliferate.db.models.billing import BillingSubject, BillingSubscription
 from proliferate.db.store.billing import (
     apply_payment_failed_hold,
@@ -75,8 +79,6 @@ from proliferate.server.billing.seats import pro_period_grant_source_ref
 from proliferate.server.billing.service import (
     activate_team_checkout_from_stripe_session,
     reconcile_initial_org_subscription_seats,
-    run_billing_store_read,
-    run_billing_store_write,
 )
 from proliferate.server.notifications import (
     BillingSlackEvent,
@@ -93,6 +95,26 @@ BILLING_SLACK_CANCELLED_STATUSES = {"canceled"}
 BILLING_SLACK_PRE_START_STATUSES = {"incomplete"}
 
 logger = logging.getLogger(__name__)
+
+
+async def _run_billing_store_read[T, **P](
+    func: Callable[Concatenate[AsyncSession, P], Awaitable[T]],
+    /,
+    *args: P.args,
+    **kwargs: P.kwargs,
+) -> T:
+    async with db_engine.async_session_factory() as db:
+        return await func(db, *args, **kwargs)
+
+
+async def _run_billing_store_write[T, **P](
+    func: Callable[Concatenate[AsyncSession, P], Awaitable[T]],
+    /,
+    *args: P.args,
+    **kwargs: P.kwargs,
+) -> T:
+    async with db_engine.async_session_factory() as db, db.begin():
+        return await func(db, *args, **kwargs)
 
 
 @dataclass(frozen=True)
@@ -141,7 +163,7 @@ async def handle_stripe_webhook(
             status_code=400,
         )
 
-    claim = await run_billing_store_write(
+    claim = await _run_billing_store_write(
         claim_webhook_event,
         provider=STRIPE_PROVIDER,
         event_id=event_id,
@@ -159,14 +181,14 @@ async def handle_stripe_webhook(
         try:
             dispatch_notifications = await _dispatch_stripe_event(event)
         except Exception as exc:
-            await run_billing_store_write(
+            await _run_billing_store_write(
                 mark_webhook_event_failed_by_id,
                 receipt_id=claim.receipt.id,
                 error=str(exc),
             )
             raise
         notifications = tuple(dispatch_notifications or ())
-        await run_billing_store_write(
+        await _run_billing_store_write(
             mark_webhook_event_processed_by_id,
             receipt_id=claim.receipt.id,
         )
@@ -225,7 +247,7 @@ async def _subject_from_object(stripe_object: dict[str, Any]) -> BillingSubject 
             billing_subject_id = UUID(subject_id)
         except ValueError:
             return None
-    return await run_billing_store_read(
+    return await _run_billing_store_read(
         get_billing_subject_for_stripe_reference,
         billing_subject_id=billing_subject_id,
         stripe_customer_id=_id_from_expandable(stripe_object.get("customer")),
@@ -257,7 +279,7 @@ async def _handle_checkout_session_completed(
     lines = await stripe_billing.list_checkout_session_line_items(session_id)
     if not stripe_billing.line_items_include_price(lines, settings.stripe_refill_10h_price_id):
         return
-    await run_billing_store_write(
+    await _run_billing_store_write(
         ensure_billing_grant_record,
         user_id=subject.user_id,
         billing_subject_id=subject.id,
@@ -294,7 +316,7 @@ async def _sync_subscription(subscription: dict[str, Any]) -> BillingSubscriptio
         monthly_item_id=monthly_item_id,
         metered_item_id=metered_item_id,
     )
-    record = await run_billing_store_write(
+    record = await _run_billing_store_write(
         upsert_stripe_subscription_record,
         billing_subject_id=subject.id,
         stripe_subscription_id=subscription_id,
@@ -417,7 +439,7 @@ async def _build_billing_slack_notification(
 async def _load_billing_subscription_by_stripe_subscription_id(
     stripe_subscription_id: str,
 ) -> BillingSubscription | None:
-    return await run_billing_store_read(
+    return await _run_billing_store_read(
         get_billing_subscription_by_stripe_subscription_id,
         stripe_subscription_id,
     )
@@ -490,7 +512,7 @@ async def _handle_invoice_paid(invoice: dict[str, Any]) -> tuple[BillingSlackNot
         and subscription_record.current_period_start is not None
     ):
         period_start_unix = int(subscription_record.current_period_start.timestamp())
-        await run_billing_store_write(
+        await _run_billing_store_write(
             ensure_billing_grant_record,
             user_id=subject.user_id,
             billing_subject_id=subject.id,
@@ -506,7 +528,7 @@ async def _handle_invoice_paid(invoice: dict[str, Any]) -> tuple[BillingSlackNot
             ),
             top_up_existing=True,
         )
-    await run_billing_store_write(clear_payment_failed_holds, billing_subject_id=subject.id)
+    await _run_billing_store_write(clear_payment_failed_holds, billing_subject_id=subject.id)
     return notifications
 
 
@@ -524,7 +546,7 @@ async def _handle_invoice_payment_failed(invoice: dict[str, Any]) -> None:
             subject = await _subject_from_object(subscription)
     if subject is None:
         return
-    await run_billing_store_write(
+    await _run_billing_store_write(
         apply_payment_failed_hold,
         billing_subject_id=subject.id,
         source=PAYMENT_HOLD_SOURCE,
@@ -536,7 +558,7 @@ async def _apply_payment_hold_for_subscription(subscription: dict[str, Any]) -> 
     subject = await _subject_from_object(subscription)
     if subject is None:
         return
-    await run_billing_store_write(
+    await _run_billing_store_write(
         apply_payment_failed_hold,
         billing_subject_id=subject.id,
         source=PAYMENT_HOLD_SOURCE,
